+++ app/src/main/java/com/battlelancer/seriesguide/lists/ManageListsDialogFragment.kt
@@ -1,38 +1,28 @@
 package com.battlelancer.seriesguide.lists
 
 import android.app.Dialog
-import android.content.Context
 import android.content.DialogInterface
-import android.database.Cursor
 import android.os.Bundle
-import android.util.SparseBooleanArray
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AdapterView
-import android.widget.Checkable
-import android.widget.CheckedTextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatDialogFragment
 import androidx.core.view.isGone
-import androidx.cursoradapter.widget.CursorAdapter
 import androidx.fragment.app.FragmentManager
-import androidx.lifecycle.lifecycleScope
-import androidx.loader.app.LoaderManager
-import androidx.loader.content.CursorLoader
-import androidx.loader.content.Loader
+import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
 import com.battlelancer.seriesguide.R
 import com.battlelancer.seriesguide.databinding.DialogManageListsBinding
+import com.battlelancer.seriesguide.databinding.ItemListCheckedBinding
+import com.battlelancer.seriesguide.lists.ManageListsDialogFragmentViewModel.ListWithItem
 import com.battlelancer.seriesguide.provider.SeriesGuideContract.ListItemTypes
-import com.battlelancer.seriesguide.provider.SeriesGuideContract.ListItems
-import com.battlelancer.seriesguide.provider.SeriesGuideContract.Lists
-import com.battlelancer.seriesguide.provider.SeriesGuideDatabase.Tables
-import com.battlelancer.seriesguide.provider.SgRoomDatabase
 import com.battlelancer.seriesguide.util.ViewTools
 import com.battlelancer.seriesguide.util.safeShow
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 
 /**
  * Displays a dialog displaying all user created lists,
@@ -41,7 +31,9 @@ import kotlinx.coroutines.withContext
 class ManageListsDialogFragment : AppCompatDialogFragment() {
 
     private var binding: DialogManageListsBinding? = null
-    private lateinit var adapter: ListsAdapter
+    private val model by viewModels<ManageListsDialogFragmentViewModel> {
+        ManageListsDialogFragmentViewModelFactory(requireActivity().application, showId)
+    }
     private var showId: Long = 0
 
     /**
@@ -59,16 +51,32 @@ class ManageListsDialogFragment : AppCompatDialogFragment() {
         this.binding = binding
 
         binding.textViewManageListsError.isGone = true
-        /*
-         * As using CHOICE_MODE_MULTIPLE does not seem to work before Jelly
-         * Bean, do everything ourselves.
-         */
-        binding.list.onItemClickListener = onItemClickListener
-        adapter = ListsAdapter(requireContext())
-        binding.list.adapter = adapter
+        val adapter = ListsWithItemAdapter(onItemClickListener)
+        binding.recyclerViewManageLists.also {
+            val layoutManager = LinearLayoutManager(it.context)
+            it.layoutManager = layoutManager
+            it.adapter = adapter
+            it.itemAnimator = null // Disable to let checkbox show its change animation.
+        }
 
-        lifecycleScope.launchWhenCreated {
-            loadShowAndListDetails()
+        // Note: viewLifeCycleOwner not available for DialogFragment, use DialogFragment itself.
+        model.showDetails.observe(this) { showDetails ->
+            if (showDetails == null) {
+                Toast.makeText(requireContext(), R.string.database_error, Toast.LENGTH_LONG).show()
+                dismiss()
+                return@observe
+            }
+            showTmdbId = showDetails.tmdbId ?: 0
+
+            binding.item.text = showDetails.title
+
+            if (showTmdbId <= 0) {
+                // Note: see OK button handler that prevents changing lists if not migrated.
+                ViewTools.configureNotMigratedWarning(binding.textViewManageListsError, true)
+            }
+        }
+        model.listsWithItem.observe(this) { lists ->
+            adapter.submitList(lists)
         }
 
         return MaterialAlertDialogBuilder(requireContext())
@@ -76,30 +84,26 @@ class ManageListsDialogFragment : AppCompatDialogFragment() {
             .setNegativeButton(android.R.string.cancel, null)
             .setPositiveButton(android.R.string.ok) { _: DialogInterface, _: Int ->
                 // Note: see show details loader that prevents loading list data if not migrated.
-                if (showTmdbId <= 0) dismiss()
+                if ((model.showDetails.value?.tmdbId ?: 0) <= 0) {
+                    dismiss()
+                    return@setPositiveButton
+                }
                 // add item to selected lists, remove from previously selected lists
-                val checkedLists = adapter.checkedPositions
                 val addToTheseLists: MutableList<String> = ArrayList()
                 val removeFromTheseLists: MutableList<String> = ArrayList()
-                for (position in 0 until adapter.count) {
-                    val listEntry = adapter.getItem(position) as Cursor
-
-                    val wasListChecked =
-                        !listEntry.getString(ListsQuery.LIST_ITEM_ID).isNullOrEmpty()
-                    val isListChecked = checkedLists[position]
-
-                    val listId = listEntry.getString(ListsQuery.LIST_ID)
-                    if (listId.isNullOrEmpty()) {
-                        continue  // skip, no id
+                val lists = model.listsWithItem.value
+                    ?: return@setPositiveButton // Do nothing, lists not loaded, yet.
+                lists
+                    .filter { it.isItemOnList != it.isItemOnListOriginal }
+                    .forEach {
+                        if (it.isItemOnListOriginal) {
+                            // remove from list
+                            removeFromTheseLists.add(it.listId)
+                        } else {
+                            // add to list
+                            addToTheseLists.add(it.listId)
+                        }
                     }
-                    if (wasListChecked && !isListChecked) {
-                        // remove from list
-                        removeFromTheseLists.add(listId)
-                    } else if (!wasListChecked && isListChecked) {
-                        // add to list
-                        addToTheseLists.add(listId)
-                    }
-                }
                 ListsTools.changeListsOfItem(
                     requireContext(), showTmdbId,
                     ListItemTypes.TMDB_SHOW, addToTheseLists, removeFromTheseLists
@@ -109,100 +113,73 @@ class ManageListsDialogFragment : AppCompatDialogFragment() {
             .create()
     }
 
-    private suspend fun loadShowAndListDetails() {
-        val showDetails = withContext(Dispatchers.IO) {
-            SgRoomDatabase.getInstance(requireContext()).sgShow2Helper().getShowMinimal(showId)
-        }
-        if (showDetails == null) {
-            Toast.makeText(requireContext(), R.string.database_error, Toast.LENGTH_LONG).show()
-            dismiss()
-            return
-        }
-        showTmdbId = showDetails.tmdbId ?: 0
-
-        val binding = binding ?: return
-        binding.item.text = showDetails.title
-
-        if (showTmdbId > 0) {
-            // Load lists status for this show.
-            LoaderManager.getInstance(this).initLoader(0, null, loaderCallbacks)
-        } else {
-            // Note: see OK button handler that prevents changing lists if not migrated.
-            ViewTools.configureNotMigratedWarning(binding.textViewManageListsError, true)
+    private val onItemClickListener = object : ListsWithItemAdapter.ItemClickListener {
+        override fun onItemClick(v: View, listId: String, isChecked: Boolean) {
+            model.setIsItemOnList(listId, !isChecked)
         }
     }
 
-    private val onItemClickListener = AdapterView.OnItemClickListener { _, view, position, _ ->
-        val checkable = view as Checkable
-        checkable.toggle()
-        adapter.setItemChecked(position, checkable.isChecked)
-    }
+    private class ListsWithItemAdapter(
+        private val itemClickListener: ItemClickListener
+    ) : ListAdapter<ListWithItem, ListsWithItemAdapter.ListWithItemViewHolder>(
+        ListWithItemDiffCallback()
+    ) {
 
-    private val loaderCallbacks = object : LoaderManager.LoaderCallbacks<Cursor> {
-        override fun onCreateLoader(id: Int, args: Bundle?): Loader<Cursor> {
-            // filter for this item, but keep other lists
-            val uri = Lists.buildListsWithListItemUri(
-                ListItems.generateListItemIdWildcard(showTmdbId, ListItemTypes.TMDB_SHOW)
-            )
-            return CursorLoader(
-                requireContext(), uri, ListsQuery.PROJECTION,
-                null, null, Lists.SORT_ORDER_THEN_NAME
-            )
+        interface ItemClickListener {
+            fun onItemClick(v: View, listId: String, isChecked: Boolean)
         }
 
-        override fun onLoadFinished(loader: Loader<Cursor>, data: Cursor) {
-            adapter.swapCursor(data)
-        }
+        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ListWithItemViewHolder =
+            ListWithItemViewHolder.inflate(parent, itemClickListener)
 
-        override fun onLoaderReset(loader: Loader<Cursor>) {
-            adapter.swapCursor(null)
-        }
-    }
+        override fun onBindViewHolder(holder: ListWithItemViewHolder, position: Int) =
+            holder.bindTo(getItem(position))
 
-    private class ListsAdapter(context: Context) : CursorAdapter(context, null, 0) {
-        private val layoutInflater: LayoutInflater = LayoutInflater.from(context)
-        val checkedPositions: SparseBooleanArray = SparseBooleanArray()
+        private class ListWithItemViewHolder(
+            private val binding: ItemListCheckedBinding,
+            itemClickListener: ItemClickListener
+        ) : RecyclerView.ViewHolder(binding.root) {
+            private var listWithItem: ListWithItem? = null
 
-        override fun bindView(view: View, context: Context, cursor: Cursor) {
-            val checkedView = view.findViewById<CheckedTextView>(android.R.id.text1)
-            checkedView.text = cursor.getString(ListsQuery.NAME)
-
-            val position = cursor.position
-
-            // prefer state set by user over database
-            val isChecked: Boolean
-            if (checkedPositions.indexOfKey(position) >= 0) {
-                // user has changed checked state, prefer it
-                isChecked = checkedPositions[position]
-            } else {
-                // otherwise prefer database state, check if item is in this list
-                val itemId = cursor.getString(ListsQuery.LIST_ITEM_ID)
-                isChecked = !itemId.isNullOrEmpty()
-                checkedPositions.put(position, isChecked)
+            init {
+                binding.checkedTextViewList.setOnClickListener { view ->
+                    listWithItem?.also {
+                        itemClickListener.onItemClick(
+                            view,
+                            it.listId,
+                            binding.checkedTextViewList.isChecked
+                        )
+                    }
+                }
             }
-            checkedView.isChecked = isChecked
+
+            fun bindTo(listWithItem: ListWithItem?) {
+                this.listWithItem = listWithItem
+                binding.checkedTextViewList.text = listWithItem?.listName ?: ""
+                binding.checkedTextViewList.isChecked = listWithItem?.isItemOnList ?: false
+            }
+
+            companion object {
+                fun inflate(parent: ViewGroup, itemClickListener: ItemClickListener) =
+                    ListWithItemViewHolder(
+                        ItemListCheckedBinding.inflate(
+                            LayoutInflater.from(parent.context),
+                            parent,
+                            false
+                        ),
+                        itemClickListener
+                    )
+            }
         }
 
-        override fun newView(context: Context, cursor: Cursor, parent: ViewGroup): View {
-            return layoutInflater.inflate(R.layout.item_list_checked, parent, false)
-        }
+        private class ListWithItemDiffCallback :
+            DiffUtil.ItemCallback<ListWithItem>() {
+            override fun areItemsTheSame(oldItem: ListWithItem, newItem: ListWithItem): Boolean =
+                oldItem.listId == newItem.listId
 
-        fun setItemChecked(position: Int, value: Boolean) {
-            checkedPositions.put(position, value)
-        }
-    }
+            override fun areContentsTheSame(oldItem: ListWithItem, newItem: ListWithItem): Boolean =
+                oldItem == newItem
 
-    interface ListsQuery {
-        companion object {
-            val PROJECTION = arrayOf(
-                Tables.LISTS + "." + Lists._ID,
-                Tables.LISTS + "." + Lists.LIST_ID,
-                Lists.NAME,
-                ListItems.LIST_ITEM_ID
-            )
-            const val LIST_ID = 1
-            const val NAME = 2
-            const val LIST_ITEM_ID = 3
         }
     }